fix and add component
