+++ src/components/App/app.module.css
@@ -4,7 +4,7 @@
   background-color: #f4f9ff;
   background-attachment: fixed;
   min-inline-size: 360px;
-  max-inline-size: 1280px;
+  max-inline-size: 1440px;
   margin: 0 auto;
   position: relative;
   min-height: 100vh;
