+++ src/components/ui/mobile-view/mobile-view.tsx
@@ -4,6 +4,8 @@ import { TasksPage } from "../../../pages/tasks-list/tasks-list";
 import { PageHeader } from "../../page-header/page-header";
 import { Outlet, useLocation } from "react-router-dom";
 import { SecondPageHeader } from "../../second-page-header/second-page-header";
+import { SecondPageTitle } from "../../second-page-title/second-page-title";
+import { FirstPageTitle } from "../../first-page-title/first-page-title";
 
 export const MobileView: FC = () => {
   const location = useLocation();
@@ -11,19 +13,25 @@ export const MobileView: FC = () => {
   return (
     <section className={styles.tasks}>
       {location.pathname === "/" ? (
-        <div className={styles.page}>
+        <>
           <PageHeader />
-          <ul className={`${styles.list} ${styles.listFirst}`}>
-            <TasksPage />
-          </ul>
-        </div>
+          <div className={styles.page}>
+            <FirstPageTitle />
+            <ul className={`${styles.list} ${styles.listFirst}`}>
+              <TasksPage />
+            </ul>
+          </div>
+        </>
       ) : (
-        <div className={styles.page}>
+        <>
           <SecondPageHeader />
-          <ul className={`${styles.list} ${styles.listFirst}`}>
-            <Outlet />
-          </ul>
-        </div>
+          <div className={styles.page}>
+            <SecondPageTitle />
+            <ul className={`${styles.list} ${styles.listFirst}`}>
+              <Outlet />
+            </ul>
+          </div>
+        </>
       )}
     </section>
   );
