+++ src/components/ui/page-header/page-header.module.css
@@ -10,63 +10,48 @@
 
 .filters {
   display: flex;
+  flex-direction: column;
   justify-content: flex-end;
   gap: 5px;
+  padding-top: 20px;
 }
 
 .filterButton {
-  padding: 2px 5px;
+  padding: 15px 5px;
   font-size: 16px;
   color: #999;
-  background-color: #e3e3e3;
-  border-radius: 0 0 10px 10px;
-  position: relative;
-  height: 25px;
+  background-color: #41414170;
+  border-radius: 10px 0 0;
+  clip-path: polygon(0 0, 100% 0, 100% 100%, 0% 80%);
 }
 
 .filterButton.active {
-  background-color: #ffebce;
-  height: 28px;
-}
-
-.title,
-.titleData {
-  font-size: 18px;
-  color: #6d6d6d;
-  font-style: italic;
-}
-
-.titleContainer {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 2px 5px;
-  border-bottom: 2px solid #ff8989;
+  background-color: #ffffff;
 }
 
 .sortDropdown {
-  position: absolute;
-  top: 0;
-  right: 30px;
-  z-index: 1;
-  display: grid;
-  grid-template-columns: repeat(1, max-content);
-  box-shadow: 0px 4px 6px rgba(0, 0, 0, 0.1);
+  display: flex;
+  flex-direction: column;
 }
 
 .sortDropdown button {
   display: flex;
   justify-content: center;
-  background-color: #ffebce;
   border: 1px solid #ffffff;
-  padding: 3px 5px;
   cursor: pointer;
-  font-size: 12px;
+  font-size: 14px;
   border-radius: calc(0.5rem - 2px);
 }
 
 @media (min-width: 768px) {
   .titleContainer {
     padding: 10px;
+  }
+}
+
+@media (max-width: 1023px) {
+  .filterButton {
+    border-radius: 0 10px 0 0;
+    clip-path: polygon(0px 0px, 100% 0px, 100% 80%, 0% 100%);
   }
 }
