+++ src/components/ui/second-page-title/second-page-title.tsx
@@ -1,0 +1,14 @@
+import { FC } from "react";
+import styles from "./second-page-title.module.css";
+
+type SecondPageTitleUIProps = {
+  title: string;
+};
+
+export const SecondPageTitleUI: FC<SecondPageTitleUIProps> = ({ title }) => {
+  return (
+    <div className={styles.titleContainer}>
+      <h3 className={styles.title}>{title}</h3>
+    </div>
+  );
+};
