--- src/pages/datas/datas.module.css
@@ -1,51 +1,0 @@
-.tasks,
-.data {
-  display: flex;
-  background-color: #6d7a9d;
-  border-radius: calc(0.5rem - 2px);
-  padding: 10px;
-  color: #000;
-  width: 50%;
-}
-
-.page,
-.pageData {
-  display: flex;
-  flex-direction: column;
-  border-radius: calc(1rem - 2px);
-  padding: 20px 10px 20px 20px;
-  width: 100%;
- /*  background: repeating-linear-gradient(#fff9f2, #fff9f2 20px, #00000000 22px), repeating-linear-gradient(90deg, #fff9f2, #fff9f2 20px, #00000000 22px); */
-  height: calc(100vh - 180px);
-  background-color: #fff;
-}
-
-/* .page {
-  padding: 20px 10px 20px 20px;
-} */
-
-.list{
-  display: flex;
-  flex-direction: column;
-  overflow-y: scroll;
-  padding-right: 10px;
-}
-
-.list::-webkit-scrollbar {
-  width: 5px;
-}
-
-.list::-webkit-scrollbar-track {
-  background: #ffffff00;
-  border-radius: calc(1rem - 2px);
-  margin: 10px 0;
-}
-
-.list::-webkit-scrollbar-thumb {
-  background: #888; 
-  border-radius: calc(1rem - 2px);
-}
-
-.list::-webkit-scrollbar-thumb:hover {
-  background: #555;
-}
